metering_point_devices, seed

+++ db/migrate/20140403173451_create_devices.rb
@@ -10,7 +10,7 @@ class CreateDevices < ActiveRecord::Migration
       t.string  :generator_type
       t.string  :shop_link
       t.string  :primary_energy
-      t.decimal :watt_peak
+      t.integer :watt_peak
       t.decimal :watt_hour_pa
       t.date    :commissioning
       t.boolean :mobile, default: false
